--- a/change/@react-native-windows-cli-3aac23ad-e95a-4369-b241-c4fea1afcb7a.json
+++ b/change/@react-native-windows-cli-3aac23ad-e95a-4369-b241-c4fea1afcb7a.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "[0.66] fix(cli): fix cache not being used when there are no dependencies",
+  "packageName": "@react-native-windows/cli",
+  "email": "jthysell@microsoft.com",
+  "dependentChangeType": "patch"
+}